Game description.md
To play the tic-tac-toe game the PC plays (X) first in the middle of the 3 × 3 grid while the player must continue from there to get a win. The player inputs (Os) using the numbers representing the positions on the board. The player wins by getting 3 consecutives Os vertically, horizontally or diagonally.